--- src/main/java/bumblebee/xchangepass/domain/transaction/rdbmsV/dto/TransactionType.java
@@ -1,5 +1,0 @@
-package bumblebee.xchangepass.domain.transaction.rdbmsV.dto;
-
-public enum TransactionType {
-    WALLET, CARD, EXCHANGE
-}
