+++ src/main/java/bumblebee/xchangepass/domain/transaction/rdbmsV/entity/TransactionType.java
@@ -1,0 +1,5 @@
+package bumblebee.xchangepass.domain.transaction.rdbmsV.entity;
+
+public enum TransactionType {
+    WALLET, CARD, EXCHANGE
+}
